compilation error fix #2
thanks for your comment,
about the compilation error I removed the comment above the constructor of Combinations. You should not have problem now.
- **-std=c++11** done
- **seriac** changed to serial
- **s value** Public and secret key with p (is a safe prime) are chosen so that gcd(p-1,s)==1.
I added a comment to main and to the function which set these values.
- **line 45** 70 was a mistake comment changed to 65 to all files accordingly
- **do you use both main.cpp and main2.cpp?**
main.cpp is for testing purposes and the difference is that it only runs one round of the attack.
I added that information at readme file.
- **parallel version** I added some info to the readme file, and to the func2.
- **Is it necessary to use OpenSSL for md5?**
No it is not. In this implementation we use md5-hashing so we found convenient to use OpenSSL's library for calculating it. You can always change that implementation another.